+++ kurtosis-module/impl/module.go
@@ -9,6 +9,7 @@ import (
 	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network/cl/teku"
 	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network/el"
 	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network/el/geth"
+	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/participant_network/el/nethermind"
 	"github.com/kurtosis-tech/eth2-merge-kurtosis-module/kurtosis-module/impl/prelaunch_data_generator"
 	"github.com/kurtosis-tech/kurtosis-core-api-lib/api/golang/lib/enclaves"
 	"github.com/kurtosis-tech/stacktrace"
@@ -16,9 +17,6 @@ import (
 	"path"
 	"text/template"
 	"time"
-
-	// "path"
-	// "text/template"
 )
 
 const (
@@ -100,6 +98,10 @@ func (e ExampleExecutableKurtosisModule) Execute(enclaveCtx *enclaves.EnclaveCon
 	if err != nil {
 		return "", stacktrace.Propagate(err, "An error occurred parsing the CL mnemonics YAML template")
 	}
+	nethermindGenesisJsonTemplate, err := parseTemplate(nethermindGenesisJsonTemplateFilepath)
+	if err != nil {
+		return "", stacktrace.Propagate(err, "An error occurred parsing the Nethermind genesis json template")
+	}
 	prelaunchData, err := prelaunch_data_generator.GeneratePrelaunchData(
 		enclaveCtx,
 		gethGenesisConfigTemplate,
@@ -121,25 +123,14 @@ func (e ExampleExecutableKurtosisModule) Execute(enclaveCtx *enclaves.EnclaveCon
 	}
 	logrus.Info("Successfully generated prelaunch data")
 
-	// TODO Nethermind template-filling here
-	/*
-	tmpl, err := template.New(templateFilename).ParseFiles(templateFilepath)
-	template.New(
-		// For some reason, the template name has to match the basename of the file:
-		//  https://stackoverflow.com/questions/49043292/error-template-is-an-incomplete-or-empty-template
-		path.Base(nethermindGenesisJsonTemplateFilepath),
-	).Parse(
-		gethGenesisJsonFilepath,
-	)
-	if err != nil {
-		return "", stacktrace.Propagate(err, "An error occurred parsing the Nethermind genesis JSON template file '%v'", nethermindGenesisJsonTemplateFilepath)
-	}
-	 */
-
 	logrus.Info("Creating EL & CL client launchers...")
 	elClientLaunchers := map[participant_network.ParticipantELClientType]el.ELClientLauncher{
 		participant_network.ParticipantELClientType_Geth: geth.NewGethELClientLauncher(
 			prelaunchData.GethELGenesisJsonFilepathOnModuleContainer,
+		),
+		participant_network.ParticipantELClientType_Nethermind: nethermind.NewNethermindELClientLauncher(
+			nethermindGenesisJsonTemplate,
+			totalTerminalDifficulty,
 		),
 	}
 	clGenesisPaths := prelaunchData.CLGenesisPaths
@@ -154,7 +145,7 @@ func (e ExampleExecutableKurtosisModule) Execute(enclaveCtx *enclaves.EnclaveCon
 	}
 	logrus.Info("Successfully created EL & CL client launchers")
 
-	logrus.Info("Adding participants...")
+	logrus.Infof("Adding %v participants...", numParticipants)
 	keystoresGenerationResult := prelaunchData.KeystoresGenerationResult
 	network := participant_network.NewParticipantNetwork(
 		enclaveCtx,
@@ -175,6 +166,7 @@ func (e ExampleExecutableKurtosisModule) Execute(enclaveCtx *enclaves.EnclaveCon
 		}
 		allClClientContexts = append(allClClientContexts, participant.GetCLClientContext())
 	}
+	logrus.Infof("Successfully added %v partitipcants", numParticipants)
 
 	logrus.Info("Launching forkmon...")
 	forkmonConfigTemplate, err := parseTemplate(forkmonConfigTemplateFilepath)